Optimise size of model structs (#2464)
This commit reduces the sizes of many integer fields, or converts data to enums when appropriate:

- Reduce the size of rate_limit fields
- Reduce the size of channel position
- Reduce size of role position
- Use an enum for `auto_archive_duration`
- `DefaultReaction` -> `ForumEmoji`
- Swap `AfkTimeout` for enum
- Reduce `Sticker::sort_value` to u16
- Reduce size of `min`/`max_values`
- Reduce size of `max_uses` and `max_age`
- Reduce size of `ActivityParty::size`
- Reduce size of `Ready::version`
- Reduce `InputText::min/max_length`
- Reduce `size`/`width`/`height` -- Discord filesize limits mean we won't hit the size limit, and therefore won't hit height/width limits, plus discord rejects 0 sized files so we can use `NonZero*`
- Reduce size of `mention_total_limit`
- Reduce size of `RoleSubscriptionData::total_months_subscribed` -- This gives us ~5000 years of role subscription
- Fix Change sizes

+++ src/model/application/component.rs
@@ -201,9 +201,9 @@ pub struct SelectMenu {
     /// The placeholder shown when nothing is selected.
     pub placeholder: Option<String>,
     /// The minimum number of selections allowed.
-    pub min_values: Option<u64>,
+    pub min_values: Option<u8>,
     /// The maximum number of selections allowed.
-    pub max_values: Option<u64>,
+    pub max_values: Option<u8>,
     /// Whether select menu is disabled.
     #[serde(default)]
     pub disabled: bool,
@@ -253,10 +253,10 @@ pub struct InputText {
     pub label: Option<String>,
     /// Minimum input length for a text input; min 0, max 4000
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub min_length: Option<u64>,
+    pub min_length: Option<u16>,
     /// Maximum input length for a text input; min 1, max 4000
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_length: Option<u64>,
+    pub max_length: Option<u16>,
     /// Whether this component is required to be filled (defaults to true)
     #[serde(default = "default_true")]
     pub required: bool,
